generalize analysis and forecasts... can handle these as a passive target
